extract code to methods
